Move dag to dag folder
